+++ lib/common/request.dart
@@ -17,28 +17,14 @@ class Request {
   String? userAgent;
 
   Request() {
-    dio = Dio(
-      BaseOptions(
-        headers: {'User-Agent': browserUa},
-        connectTimeout: const Duration(seconds: 10),
-        receiveTimeout: const Duration(seconds: 15),
-        sendTimeout: const Duration(seconds: 10),
-      ),
-    );
-    _clashDio = Dio(
-      BaseOptions(
-        connectTimeout: const Duration(seconds: 10),
-        receiveTimeout: const Duration(seconds: 60),
-        sendTimeout: const Duration(seconds: 10),
-      ),
-    );
+    dio = Dio(BaseOptions(headers: {'User-Agent': browserUa}));
+    _clashDio = Dio();
     _clashDio.httpClientAdapter = IOHttpClientAdapter(
       createHttpClient: () {
         final client = HttpClient();
-        client.idleTimeout = const Duration(seconds: 30);
         client.findProxy = (Uri uri) {
           client.userAgent = globalState.ua;
-          return PSGHttpOverrides.handleFindProxy(uri);
+          return FlClashHttpOverrides.handleFindProxy(uri);
         };
         return client;
       },